+++ src/main/java/com/clova/anifriends/domain/volunteer/dto/response/FindVolunteerMyPageResponse.java
@@ -1,6 +1,6 @@
 package com.clova.anifriends.domain.volunteer.dto.response;
 
-import com.clova.anifriends.domain.applicant.wrapper.ApplicantStatus;
+import com.clova.anifriends.domain.applicant.vo.ApplicantStatus;
 import com.clova.anifriends.domain.volunteer.Volunteer;
 import com.clova.anifriends.domain.volunteer.vo.VolunteerGender;
 import java.time.LocalDate;
