+++ src/test/java/com/clova/anifriends/docs/enumtype/EnumDocsController.java
@@ -3,7 +3,7 @@ package com.clova.anifriends.docs.enumtype;
 import com.clova.anifriends.domain.animal.vo.AnimalActive;
 import com.clova.anifriends.domain.animal.vo.AnimalGender;
 import com.clova.anifriends.domain.animal.vo.AnimalType;
-import com.clova.anifriends.domain.applicant.wrapper.ApplicantStatus;
+import com.clova.anifriends.domain.applicant.vo.ApplicantStatus;
 import com.clova.anifriends.domain.common.EnumType;
 import com.clova.anifriends.domain.volunteer.vo.VolunteerGender;
 import java.util.Arrays;
