+++ src/test/java/com/clova/anifriends/domain/applicant/repository/ApplicantRepositoryTest.java
@@ -1,10 +1,10 @@
 package com.clova.anifriends.domain.applicant.repository;
 
 import static com.clova.anifriends.domain.applicant.support.ApplicantFixture.applicant;
-import static com.clova.anifriends.domain.applicant.wrapper.ApplicantStatus.ATTENDANCE;
-import static com.clova.anifriends.domain.applicant.wrapper.ApplicantStatus.NO_SHOW;
-import static com.clova.anifriends.domain.applicant.wrapper.ApplicantStatus.PENDING;
-import static com.clova.anifriends.domain.applicant.wrapper.ApplicantStatus.REFUSED;
+import static com.clova.anifriends.domain.applicant.vo.ApplicantStatus.ATTENDANCE;
+import static com.clova.anifriends.domain.applicant.vo.ApplicantStatus.NO_SHOW;
+import static com.clova.anifriends.domain.applicant.vo.ApplicantStatus.PENDING;
+import static com.clova.anifriends.domain.applicant.vo.ApplicantStatus.REFUSED;
 import static com.clova.anifriends.domain.recruitment.support.fixture.RecruitmentFixture.recruitment;
 import static com.clova.anifriends.domain.shelter.support.ShelterFixture.shelter;
 import static com.clova.anifriends.domain.volunteer.support.VolunteerFixture.volunteer;
